implemented import tests for ddahs; all import tests finished
