+++ frontend/src/tests/import-export-data/import-data.js
@@ -2,10 +2,12 @@ import instructorsJSON from "./instructors.json";
 import applicantsJson from "./applicants.json";
 import positionsJson from "./positions.json";
 import assignmentsJson from "./assignments.json";
+import ddahsJson from "./ddahs.json";
 
 export const objectJSON = {
     instructors: instructorsJSON["instructors"],
     applicants: applicantsJson["applicants"],
     positions: positionsJson["positions"],
     assignments: assignmentsJson["assignments"],
+    ddahs: ddahsJson["ddahs"],
 };
